+++ steelc/src/parser/ast/ast_visitor.cpp
@@ -79,6 +79,12 @@ void ast_visitor::visit(std::shared_ptr<initializer_list> init) {
 	}
 }
 void ast_visitor::visit(std::shared_ptr<function_call> func_call) {
+	// accept callee for methods
+	if (func_call->is_method()) {
+		func_call->callee->accept(*this);
+	}
+
+	// accept all args
 	for (auto& arg : func_call->args) {
 		arg->accept(*this);
 	}