+++ steelc/src/parser/semantics/type_checker.h
@@ -21,6 +21,7 @@ public:
 	void visit(std::shared_ptr<unary_expression> expr) override;
 	void visit(std::shared_ptr<index_expression> expr) override;
 	void visit(std::shared_ptr<cast_expression> expr) override;
+	void visit(std::shared_ptr<member_expression> expr) override;
 	void visit(std::shared_ptr<function_call> func_call) override;
 	void visit(std::shared_ptr<if_statement> if_stmt) override;
 	void visit(std::shared_ptr<inline_if> inline_if) override;